+++ src/services/data-fetch/admin-page-data-fetch/adminPageDataFetch.ts
@@ -1,3 +1,3 @@
-import { adminPageMock } from "../../../utils/mockData/admin-page/adminPage";
+import { adminPageMock } from "../../../utils/mock-data/admin-page/adminPage";
 
 export const adminPageDataFetch = async () => adminPageMock;
