+++ src/services/data-fetch/user-pages-data-fetch/home-page-data-fetch/homePageDataFetch.ts
@@ -1,3 +1,3 @@
-import { homePageMock } from "../../../../utils/mockData/user-pages/home-page/home-page";
+import { homePageMock } from "../../../../utils/mock-data/user-pages/home-page/home-page";
 
 export const homePageDataFetch = async () => homePageMock;
