+++ src/services/data-fetch/user-pages-data-fetch/page-1-data-fetch/page1DataFetch.ts
@@ -1,3 +1,3 @@
-import { page1Mock } from '../../../../utils/mockData/user-pages/page-1/page1';
+import { page1Mock } from '../../../../utils/mock-data/user-pages/page-1/page1';
 
 export const page1DataFetch = async () => page1Mock;
