+++ src/services/data-fetch/user-pages-data-fetch/page-2-data-fetch/page2DataFetch.ts
@@ -1,3 +1,3 @@
-import { page2Mock } from '../../../../utils/mockData/user-pages/page-2/page2';
+import { page2Mock } from '../../../../utils/mock-data/user-pages/page-2/page2';
 
 export const page2DataFetch = async () => page2Mock;
