+++ src/utils/mock-data/user-pages/home-page/home-page.ts
@@ -1,4 +1,4 @@
 export const homePageMock = {
     header: "Home page",
     content: "Here is home page"
-}
+};
